+++ lib/ibm/cloud/sdk/vpc/http/vpc_collection.rb
