+++ lib/ibm/cloud/sdk/vpc/http/vpc_http.rb
