+++ lib/ibm/cloud/sdk/vpc/http/vpc_instance.rb
